+++ lib/src/state_builder.dart
@@ -28,6 +28,7 @@ class StateBuilder<T> extends StatefulWidget {
     this.afterInitialBuild,
     this.afterRebuild,
     this.disposeModels,
+    this.watch,
   })  : assert(builder != null || builderWithChild != null, '''
   
   | ***Builder not defined***
@@ -169,6 +170,18 @@ class StateBuilder<T> extends StatefulWidget {
   ///In any of the injected class you can define a 'dispose()' method to clean up resources.
   final bool disposeModels;
 
+  ///A function that returns a one instance variable or a list of
+  ///them. The rebuild process will be triggered if at least one of
+  ///the return variable changes.
+  ///
+  ///Return variable must be either a primitive variable, a List, a Map or a Set.
+  ///
+  ///To use a custom type, you should override the `toString` method to reflect
+  ///a unique identity of each instance.
+  ///
+  ///If it is not defined all listener will be notified when a new state is available.
+  final Object Function(ReactiveModel<T> model) watch;
+
   @override
   _StateBuilderState<T> createState() => _StateBuilderState<T>();
 }
@@ -241,34 +254,48 @@ To fix, you have to either :
         (_) => widget.afterInitialBuild(context, _exposedModel),
       );
     }
+
+    if (widget.watch != null) {
+      _after = widget.watch(_exposedModel)?.toString();
+    }
   }
 
+  String _before;
+  String _after;
   @override
   bool update([void Function(BuildContext) onSetState]) {
     if (!mounted) {
       return false;
     }
+    if (widget.watch != null) {
+      _before = widget.watch(_exposedModel)?.toString();
+    }
 
-    setState(
-      () {
-        if (onSetState != null) {
-          onSetState(context);
-        }
+    if (widget.watch == null ||
+        _before.hashCode != _after.hashCode ||
+        !identical(_before, _after)) {
+      setState(
+        () {
+          if (onSetState != null) {
+            onSetState(context);
+          }
 
-        //Do not call [StateBuilder.onSetState] more than one for each rebuild
-        if (!_isDirty) {
-          if (widget.onSetState != null) {
-            widget.onSetState(context, _exposedModel);
+          //Do not call [StateBuilder.onSetState] more than one for each rebuild
+          if (!_isDirty) {
+            if (widget.onSetState != null) {
+              widget.onSetState(context, _exposedModel);
+            }
+            if (widget.onRebuildState != null) {
+              WidgetsBinding.instance.addPostFrameCallback(
+                (_) => widget.onRebuildState(context, _exposedModel),
+              );
+            }
+            _isDirty = true;
           }
-          if (widget.onRebuildState != null) {
-            WidgetsBinding.instance.addPostFrameCallback(
-              (_) => widget.onRebuildState(context, _exposedModel),
-            );
-          }
-          _isDirty = true;
-        }
-      },
-    );
+        },
+      );
+    }
+    _after = _before;
     return true;
   }